Move FailingSymbolServer to test.rs
